% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.assign.sharing.settings.R
\name{ds.assign.sharing.settings}
\alias{ds.assign.sharing.settings}
\title{assigns settings on each data servers.}
\usage{
ds.assign.sharing.settings(datasources = NULL)
}
\arguments{
\item{datasources}{- a list of \code{DSConnection-class} objects obtained after login.}
}
\description{
assigns the initial settings required for sharing parametres or
encoded information.
}
