% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.read.R
\name{ds.read}
\alias{ds.read}
\title{retrieve encrypted data from datashield servers}
\usage{
ds.read(
  data.from.server = NULL,
  data.encrypted = NULL,
  data.held.in.server = "D",
  no.rows = 1000,
  client.side.variable = NULL,
  datasources = NULL
)
}
\arguments{
\item{data.from.server}{a list of encrypted data obtained from some dataSHIELD server}

\item{data.encrypted}{a character variable representing the name of the R object of encrypted data}

\item{data.held.in.server}{a character variable representing the name of the R object of the data on}

\item{no.rows}{a numerical variable representing to be transferred at each iteration.}

\item{client.side.variable}{a character variable representing the name of an R object}

\item{datasources}{a list of connections to dataSHIELD servers}
}
\description{
retrieve encrypted data from datashield servers
}
