% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.write.R
\name{ds.write}
\alias{ds.write}
\title{write outcome of computations to some DataSHIELD server}
\usage{
ds.write(
  data.to.server = NULL,
  class.type.server = "data.frame",
  no.rows = 1000,
  client.side.variable = NULL,
  column.server = NULL,
  datasources = NULL
)
}
\arguments{
\item{data.to.server}{- a character argument representing the name of new server variable}

\item{no.rows}{- a numerical (integer) argument indicatind the maximum of rows used
in one transfer of data}

\item{client.side.variable}{- a character argument representing the name of client.side.variable
storing the outcome of the computations}

\item{column.server}{- a character argument representing the name of column used to indicate
the server origin}

\item{datasources}{a list of connections to dataSHIELD servers}
}
\description{
. This client-side function writes some data to one or more DataSHIELD servers. The
function aims to sent some outcome of computations used in learning algorithms or situations where
some servers needs storing outcomes of computations.

It  is inspired from the federated learning model in this publication [Federated Machine Learning: Concept
and applications](https://arxiv.org/pdf/1902.04885.pdf).
}
