% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.transfer.R
\name{dsce.check.data.encrypted}
\alias{dsce.check.data.encrypted}
\title{check the data has been suitably encrypted}
\usage{
dsce.check.data.encrypted(
  data.server = NULL,
  data.encrypted = NULL,
  data.held.in.server = "D",
  datasources = NULL
)
}
\value{
TRUE data are apprpriately encrypted on every DataSHIELD server. Otherwise, FALSE
}
\description{
call the server-side function \code{isDataEncodedDS} from the server package dsShareServer
}
