% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.share.param.R
\name{dssp.assign.param.settings}
\alias{dssp.assign.param.settings}
\title{assigns on a datashield server a parameter settings.}
\usage{
dssp.assign.param.settings(connection, param.names = NULL)
}
\arguments{
\item{connection}{- datasources}

\item{param.names}{- param names}
}
\description{
assigns on a datashield server a parameter settings.
}
