% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.share.param.R
\name{dssp.assign.settings}
\alias{dssp.assign.settings}
\title{assigns settings on each data servers.}
\usage{
dssp.assign.settings(connections)
}
\arguments{
\item{connections}{- datasources}
}
\description{
assigns settings on each data servers.
}
