% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.share.param.R
\name{dssp.complete.exchange}
\alias{dssp.complete.exchange}
\title{launch the steps a whole exchange. Most servers become a master.}
\usage{
dssp.complete.exchange(connections, param.names = NULL, tolerance = 15)
}
\arguments{
\item{connections}{- datasources}

\item{param.names}{- param names}

\item{tolerance}{- decimal places}
}
\description{
launch the steps a whole exchange. Most servers become a master.
}
