% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.share.param.R
\name{dssp.decrypt.param}
\alias{dssp.decrypt.param}
\title{decrypt parameter}
\usage{
dssp.decrypt.param(connection, param.names, tolerance = 15)
}
\arguments{
\item{connection}{- datasources}

\item{param.names}{- param names}
}
\description{
decrypt parameter
}
