% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.share.param.R
\name{dssp.encrypt.data}
\alias{dssp.encrypt.data}
\title{encrypt data}
\usage{
dssp.encrypt.data(connection, master_mode = TRUE, preserve_mode = FALSE)
}
\arguments{
\item{connection}{- datasources}

\item{master_mode}{- boolean}

\item{preserve_mode}{- boolean}
}
\description{
encrypt data
}
