% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.share.param.R
\name{dssp.exchange}
\alias{dssp.exchange}
\title{complete an exchange between a master and a receiver. Both of these argument is a single
connection to a server}
\usage{
dssp.exchange(master, receiver, param.names = NULL, tolerance = 15)
}
\arguments{
\item{master}{- datasource}

\item{receiver}{- datasource}

\item{param.names}{- param names}

\item{tolerance}{- decimal places}
}
\description{
complete an exchange between a master and a receiver. Both of these argument is a single
connection to a server
}
