% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.share.param.R
\name{dssp.remove.encryption.data}
\alias{dssp.remove.encryption.data}
\title{delete variable used to encrypt data on a server}
\usage{
dssp.remove.encryption.data(connection = NULL, master.mode = TRUE)
}
\arguments{
\item{connection}{- datasource}

\item{master.mode}{- boolean}
}
\description{
delete variable used to encrypt data on a server
}
