% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.share.param.R
\name{dssp.remove.exchange.data}
\alias{dssp.remove.exchange.data}
\title{delete from each servers}
\usage{
dssp.remove.exchange.data(connections)
}
\arguments{
\item{connections}{- datasources}
}
\description{
delete from each servers
}
