% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.share.param.R
\name{dssp.remove.existing.parameters}
\alias{dssp.remove.existing.parameters}
\title{delete from servers parameters}
\usage{
dssp.remove.existing.parameters(connections, param.names = NULL)
}
\arguments{
\item{connections}{- datasources}

\item{param.names}{- param names}
}
\description{
delete from servers parameters
}
