% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.share.param.R
\name{dssp.share.parameter}
\alias{dssp.share.parameter}
\title{Helper function that completes the sharing process}
\usage{
dssp.share.parameter(param.names = NULL, tolerance = 15, datasources = NULL)
}
\arguments{
\item{param.names}{- param names}

\item{tolerance}{- decimal places}

\item{datasources}{- connections}
}
\description{
Helper function that completes the sharing process
}
