% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.share.param.R
\name{dssp.transfer.coordinates}
\alias{dssp.transfer.coordinates}
\title{transform coordinates}
\usage{
dssp.transfer.coordinates(sender = NULL, receiver = NULL)
}
\arguments{
\item{sender}{- datasource}

\item{receiver}{- datasource}
}
\description{
transform coordinates
}
