% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.share.param.R
\name{dssp.transfer.encrypted.matrix}
\alias{dssp.transfer.encrypted.matrix}
\title{transfer encrypted matrices from one dataSHIELD server to another}
\usage{
dssp.transfer.encrypted.matrix(
  sender = NULL,
  receiver = NULL,
  master_mode = TRUE
)
}
\arguments{
\item{sender}{- datasource}

\item{receiver}{- datasource}

\item{master_mode}{- boolean}
}
\description{
transfer encrypted matrices from one dataSHIELD server to another
}
