% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.share.param.R
\name{dssp.transform.outcome.to.logical}
\alias{dssp.transform.outcome.to.logical}
\title{single or multiple values are transformed in one logical values.}
\usage{
dssp.transform.outcome.to.logical(value)
}
\arguments{
\item{value}{value}
}
\description{
single or multiple values are transformed in one logical values.
}
