% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.read.R
\name{dstr.check.data.encrypted}
\alias{dstr.check.data.encrypted}
\title{check the data has been suitably encrypted}
\usage{
dstr.check.data.encrypted(
  data.server = NULL,
  data.encrypted = NULL,
  data.held.in.server = "D",
  datasources = NULL
)
}
\arguments{
\item{data.encrypted}{a character variable representing the name of the R object of encrypted data}

\item{data.held.in.server}{a character variable representing the name of the R object of the data on}

\item{datasources}{a list of connections to dataSHIELD servers}

\item{data.from.server}{a list of encrypted data obtained from some dataSHIELD server}

\item{client.side.variable}{a character variable representing the name of an R object}
}
\value{
TRUE data are apprpriately encrypted on every DataSHIELD server. Otherwise, FALSE
}
\description{
call the server-side function \code{isDataEncodedDS} from the server package dsShareServer
}
