% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.transfer.R
\name{dstr.check.param}
\alias{dstr.check.param}
\title{check arguments are correct}
\usage{
dstr.check.param(
  data.server = NULL,
  data.encrypted = NULL,
  data.held.in.server = "D",
  datasources = NULL
)
}
\value{
TRUE - all correct data type. FALSE - if checks have passed, but a class type is not correct.
}
\description{
throw some errors or return a logical value
}
\note{
throws errors CLIENT:SHARE:ERR:100 to CLIENT:SHARE:ERR:103
}
