% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.read.R
\name{dstr.concatenate}
\alias{dstr.concatenate}
\title{concatenate encrypted data to a client-side dataframe}
\usage{
dstr.concatenate(data.from.server = list(), client.side.variable = NULL)
}
\arguments{
\item{data.from.server}{a list of encrypted data obtained from some dataSHIELD server}

\item{client.side.variable}{a character variable representing the name of an R object
in the global environment.}
}
\description{
concatenate encrypted data to a client-side dataframe
}
