% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.read.R
\name{dstr.extract.encrypted.data}
\alias{dstr.extract.encrypted.data}
\title{extract data from encrypted and encoded data.}
\usage{
dstr.extract.encrypted.data(data.from.server = list())
}
\arguments{
\item{data.from.server}{a list of encrypted data obtained from some dataSHIELD server}
}
\value{
extracted data in a dataframe
in the global environment.
}
\description{
extract data from encrypted and encoded data.
}
