% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.read.R
\name{dstr.get.data.from.server}
\alias{dstr.get.data.from.server}
\title{read iteratively the data from the server}
\usage{
dstr.get.data.from.server(
  data.encrypted = NULL,
  no.rows = 1000,
  client.side.variable = NULL,
  datasources = NULL
)
}
\arguments{
\item{data.encrypted}{a character variable representing the name of the R object of encrypted data}

\item{no.rows}{a numerical variable representing to be transferred at each iteration.}

\item{client.side.variable}{a character variable representing the name of an R object}
}
\description{
read iteratively the data from the server
}
