% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.transfer.R
\name{dstr.is.eof}
\alias{dstr.is.eof}
\title{indicates some encrypted data remains to be transferred.}
\usage{
dstr.is.eof(data.encrypted = "", datasources = NULL)
}
\value{
TRUE if all the data have been transferred in every server
}
\description{
indicates some encrypted data remains to be transferred.
}
\note{
Server errors thrown SERVER::ERR::SHARING::001, SERVER:ERR:021, SERVER:ERR:009
It is assumed the parameter is correct and have been checked with
\code{dstr.check.param}
}
