% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.read.R
\name{dstr.next}
\alias{dstr.next}
\title{transfers encrypted data from the servers for computations}
\usage{
dstr.next(data.encrypted = "", no.rows = 10, datasources = NULL)
}
\arguments{
\item{data.encrypted}{a character variable representing the name of the R object of encrypted data}

\item{no.rows}{a numerical variable representing to be transferred at each iteration.}

\item{datasources}{a list of connections to dataSHIELD servers}
}
\value{
TRUE if all the data have been transferred without any issues.
}
\description{
transfers encrypted data from the servers for computations
}
\note{
Server errors thrown SERVER::ERR::SHARING::001, SERVER:ERR:002, SERVER:ERR:004, SERVER:ERR:021
It is assumed the parameter is correct and have been checked with \code{dstr.check.param}. It is also assumed
the data have been checked for encryption first with the function \code{dsce.check.data.encrypted}
}
