% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.write.R
\name{dswr.split}
\alias{dswr.split}
\title{split a data frame into a subsets}
\usage{
dswr.split(
  client.side.variable = "",
  column = "",
  client.side.split = "",
  datasources = NULL
)
}
\arguments{
\item{client.side.variable}{- a character argument representing the name of client.side.variable
storing the outcome of the computations}
}
\description{
This function split a data frame into subsets; one for each server.
}
