% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.assign.sharing.settings.R
\name{ds.assign.sharing.settings}
\alias{ds.assign.sharing.settings}
\title{assigns settings on each data servers.}
\usage{
ds.assign.sharing.settings(connections = NULL)
}
\arguments{
\item{connections}{- datasources}
}
\description{
assigns settings on each data servers.
}
