% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.read.R
\name{dstr.check.param}
\alias{dstr.check.param}
\title{check arguments are correct}
\usage{
dstr.check.param(
  data.server = NULL,
  data.encrypted = NULL,
  data.held.in.server = "D",
  no.rows = 1000,
  client.side.variable = NULL,
  datasources = NULL
)
}
\arguments{
\item{data.server}{a list of encrypted data obtained from some dataSHIELD server}

\item{data.encrypted}{a character variable representing the name of the R object of encrypted data}

\item{data.held.in.server}{a character variable representing the name of the R object of the data on}

\item{no.rows}{a numerical variable representing to be transferred at each iteration.}

\item{client.side.variable}{a character variable representing the name of an R object}

\item{datasources}{a list of connections to dataSHIELD servers}
}
\value{
TRUE - all correct data type. FALSE - if checks have passed, but a class type is not correct.
}
\description{
throw some errors or return a logical value
}
\note{
throws errors CLIENT:SHARE:ERR:100 to CLIENT:SHARE:ERR:103
}
